add to issues from stream
