+++ client/app/hearings/components/HearingWorksheetIssues.jsx
@@ -1,6 +1,9 @@
 import React, { PureComponent } from 'react';
 import Table from '../../components/Table';
 import PropTypes from 'prop-types';
+import Button from '../../components/Button';
+import { bindActionCreators } from 'redux';
+import { onAddIssue } from '../actions/Issue';
 import { connect } from 'react-redux';
 import HearingWorksheetIssueFields from './HearingWorksheetIssueFields';
 import HearingWorksheetPreImpressions from './HearingWorksheetPreImpressions';
@@ -9,6 +12,8 @@ import HearingWorksheetIssueDelete from './HearingWorksheetIssueDelete';
 class HearingWorksheetIssues extends PureComponent {
 
   getKeyForRow = (index) => index;
+
+  onAddIssue = (appealKey) => () => this.props.onAddIssue(appealKey);
 
   render() {
     let {
@@ -114,6 +119,12 @@ class HearingWorksheetIssues extends PureComponent {
               summary={'Worksheet Issues'}
               getKeyForRow={this.getKeyForRow}
           />
+          <Button
+                classNames={['usa-button-outline', 'hearings-add-issue']}
+                name="+ Add Issue"
+                id={`button-addIssue-${appealKey}`}
+                onClick={this.onAddIssue(appealKey)}
+              />
         </div>;
   }
 }
@@ -122,9 +133,13 @@ const mapStateToProps = (state) => ({
   HearingWorksheetIssues: state
 });
 
+const mapDispatchToProps = (dispatch) => bindActionCreators({
+  onAddIssue
+}, dispatch);
 
 export default connect(
   mapStateToProps,
+  mapDispatchToProps
 )(HearingWorksheetIssues);
 
 HearingWorksheetIssues.propTypes = {
